+++ src/data_structures/graph.jl
@@ -52,7 +52,7 @@ struct Graph{I}
         G = UndirectedGraph{I}()
         return new{I}(G)
     end
-    Graph() = Graph{Int64}()
+    Graph() = Graph{Int}()
     Graph(G::UndirectedGraph{I}) where {I} = new{I}(G)
 end
 Base.:(==)(G::Graph, H::Graph) = get_graph(G) == get_graph(H)
